Final-202010/Final-202010/src/RecursionProblems.java


public class RecursionProblems {
	
	
	/**
	  * For this problem, you are to use recursion to find if each element
	  * of one array appears in the other array in the same order.  There
	  * may be extra values in the input array.
	  * 
	  * You'll want a helper function to solve this problem
	  * 
	  * Example 1:
	  * input = {1, 2, 3, 4, 5, 6, 7}
	  * toFind = {3, 4, 5}
	  * Given the above input, this should return true
	  * 
	  * Example 2:
	  * input = {1, 2, 3, 4 ,5, 6, 7}
	  * toFind = {1, 3, 5}
	  * Given the above input, this should return true
	  * 
	  * Example 3:
	  * input = {1, 3, 5, 7, 9}
	  * toFind = {3, 4, 5}
	  * false because 4 is not in the array this should return false
	  * 
	  * Example 4:
	  * input = {1, 3, 5, 7, 4}
	  * toFind = {3, 4, 5}
	  * false because the order is 3,5,4 in the input array
	  * 
	  * @param input - the array to find the values in
	  * @param toFind - the array to find in the other
	  * @return true if toFind is found in input, false otherwise
	  */
	
	static boolean findIn(int[] input, int[] toFind) {
		if (toFind.length == 1) return true;
		return findHelper(input, toFind, 0);
//		return false;
	}
	static boolean findHelper(int[] input, int[] toFind, int ind) {
		if (ind < toFind.length) {
			return findHelper(input, toFind, ind--);
		}
		return false;
	}

}
